Move response mapping class CIRC-362
