+++ src/main/java/org/folio/circulation/domain/LoanRepository.java
@@ -4,8 +4,8 @@ import static java.util.Objects.nonNull;
 import static java.util.concurrent.CompletableFuture.completedFuture;
 import static org.folio.circulation.support.CqlQuery.exactMatch;
 import static org.folio.circulation.support.CqlQuery.exactMatchAny;
-import static org.folio.circulation.support.ResponseMapping.forwardOnFailure;
-import static org.folio.circulation.support.ResponseMapping.usingJson;
+import static org.folio.circulation.support.http.ResponseMapping.forwardOnFailure;
+import static org.folio.circulation.support.http.ResponseMapping.usingJson;
 import static org.folio.circulation.support.Result.failed;
 import static org.folio.circulation.support.Result.of;
 import static org.folio.circulation.support.Result.succeeded;
