+++ src/main/java/org/folio/circulation/support/http/ResponseMapping.java
@@ -1,10 +1,12 @@
-package org.folio.circulation.support;
+package org.folio.circulation.support.http;
 
 import static org.folio.circulation.support.Result.failed;
 import static org.folio.circulation.support.Result.of;
 
 import java.util.function.Function;
 
+import org.folio.circulation.support.ForwardOnFailure;
+import org.folio.circulation.support.Result;
 import org.folio.circulation.support.http.client.Response;
 
 import io.vertx.core.json.JsonObject;
